исправления по комментам и бонус задачи
