+++ README.md
@@ -12,41 +12,34 @@ go test -run ^TestDelTask$ ./tests
 
 2) Директория `web` содержит файлы фронтенда.
 
-3) Database.go инициализирует и создает БД. 
+3) database.go инициализирует и создает БД. Содержит всю логику и функции для работы с БД.
 
 4) auth.go отвечает за авторизацию и всех задач связанных с ней.
 
-5) Пакет task_service отвечает за реализацию функционала работы с задачами : 
-5.1) add_tasks.go - отвечает за добавление задачи в БД
+5) tasksHandlers отвечает за реализацию функционала работы с задачами : 
+добавление задачи в БД
+удаление задачи из БД
+чтения задач из БД.
+функции отметки задачи выполненной.
+обновление данных существующей задачи
 
-5.2) delete_task.go - отвечает за удаление задачи из БД
+6) rulesRepeat.go - отвечает за функционирование правил повторения и переноса задач согласно правилам :
 
-5.3) func_helpers.go - хранит вспомогательные функции (такие как подсчет високосного года, проверку срезов и тд ), а также константы и структуры для функционирования работы с задачами.
-
-5.4) get_Tasks.go - отвечает за функционал чтения задач из БД.
-
-5.5) makeDone_task.go - отвечает за реализацию функции отметки задачи выполненной.
-
-5.6) rulesRepeat_task.go - отвечает за функционирование правил повторения и переноса задач согласно правилам :
-
-5.6.1) Если правило не указано, отмеченная выполненной задача будет удаляться из таблицы;
+6.1) Если правило не указано, отмеченная выполненной задача будет удаляться из таблицы;
 d <число> — задача переносится на указанное число дней. Максимально допустимое число равно 400. Примеры:
 d 1 — каждый день;
 d 7 — для вычисления следующей даты добавляем семь дней;
 d 60 — переносим на 60 дней.
-5.6.2) y — задача выполняется ежегодно. Этот параметр не требует дополнительных уточнений. При выполнении задачи дата перенесётся на год вперёд.
-5.6.3) w <через запятую от 1 до 7> — задача назначается в указанные дни недели, где 1 — понедельник, 7 — воскресенье. Например:
+6.2) y — задача выполняется ежегодно. Этот параметр не требует дополнительных уточнений. При выполнении задачи дата перенесётся на год вперёд.
+6.3) w <через запятую от 1 до 7> — задача назначается в указанные дни недели, где 1 — понедельник, 7 — воскресенье. Например:
 w 7 — задача перенесётся на ближайшее воскресенье;
 w 1,4,5 — задача перенесётся на ближайший понедельник, четверг или пятницу;
-5.6.4) m <через запятую от 1 до 31,-1,-2> [через запятую от 1 до 12] — задача назначается в указанные дни месяца. При этом вторая последовательность чисел опциональна и указывает на определённые месяцы. Например:
+6.4) m <через запятую от 1 до 31,-1,-2> [через запятую от 1 до 12] — задача назначается в указанные дни месяца. При этом вторая последовательность чисел опциональна и указывает на определённые месяцы. Например:
 m 4 — задача назначается на четвёртое число каждого месяца;
 m 1,15,25 — задача назначается на 1-е, 15-е и 25-е число каждого месяца;
 m -1 — задача назначается на последний день месяца;
 
-6) updete_task.go - Отчечает за обновление данных существующей задачи
-
-
-Пример проверки работы авторизации: 
+7) Пример проверки работы авторизации: 
 $ export TODO_PASSWORD="123"
 $ go run .
 $ curl -X POST http://localhost:7540/api/signin \
@@ -54,7 +47,7 @@ $ curl -X POST http://localhost:7540/api/signin \
 ype: application/json"      -d '{"password":"1234"}'
 {"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InVzZXIiLCJleHAiOjE3Mjg2MTc0MTd9.ZkCib0eTSAxkX80CuL28eAsHmAwRZ4b3bSDkYdVwZn4"}
 
-Сборка и запуск докера:
-1)docker build --tag my_app:v1 . 
-2)docker run -d -p 7540:7540 my_app:v1
-3)Пароль для фронта - 1234 (Выбран как эталон кибербезопасности)
+8) Сборка и запуск докера:
+8.1)docker build --tag my_app:v1 . 
+8.2)docker run -d -p 7540:7540 my_app:v1
+8.3)Пароль для фронта - 1234 (Выбран как эталон кибербезопасности)
