+++ main.go
@@ -6,25 +6,28 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/LTVgreater5CPi/go_final_project/tasks_service"
-
 	_ "github.com/mattn/go-sqlite3"
 )
+
+// ВСПОМОГАТЕЛЬНЫЕ ФУНКЦИИ
+func MakeHandler(fn func(http.ResponseWriter, *http.Request, *sql.DB), db *sql.DB) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		fn(w, r, db)
+	}
+}
 
 func main() {
 	port := os.Getenv("TODO_PORT")
 	if port == "" {
 		port = "7540"
 	}
-
 	appPassword = os.Getenv("TODO_PASSWORD")
 	if appPassword == "" {
-		log.Println("Переменная TODO_PASSWORD не установлена. Аутентификация отключена.")
+		log.Println("The TODO_PASSWORD variable is not set. Authentication is disabled")
 	}
-
 	db, err := setupDB()
 	if err != nil {
-		log.Fatalf("Ошибка настройки БД: %v", err)
+		log.Fatalf("DB configuration error: %v", err)
 	}
 	defer db.Close()
 
@@ -33,29 +36,14 @@ func main() {
 	http.Handle("/", fileServer)
 
 	// API рабы с аутентификацией
-	http.HandleFunc("/api/nextdate", tasks_service.NextDateHandler)
-	http.HandleFunc("/api/task", authMidW(tasks_service.MakeHandler(taskHandler, db)))
-	http.HandleFunc("/api/tasks", authMidW(tasks_service.MakeHandler(tasks_service.TasksHandler, db)))
-	http.HandleFunc("/api/task/done", authMidW(tasks_service.MakeHandler(tasks_service.TaskDoneHandler, db)))
-	http.HandleFunc("/api/signin", tasks_service.MakeHandler(signInHandler, db))
+	http.HandleFunc("/api/nextdate", NextDateHandler)
+	http.HandleFunc("/api/task", authMidW(MakeHandler(taskHandler, db)))
+	http.HandleFunc("/api/tasks", authMidW(MakeHandler(GetTaskHandler, db)))
+	http.HandleFunc("/api/task/done", authMidW(MakeHandler(TaskDoneHandler, db)))
+	http.HandleFunc("/api/signin", MakeHandler(signInHandler, db))
 
+	log.Printf("Starting the server on the port %s...\n", port)
 	if err := http.ListenAndServe(":"+port, nil); err != nil {
 		log.Fatal(err)
-	}
-}
-
-// Обработчик для /api/task
-func taskHandler(w http.ResponseWriter, r *http.Request, db *sql.DB) {
-	switch r.Method {
-	case http.MethodPost:
-		tasks_service.AddTaskHandler(w, r, db)
-	case http.MethodGet:
-		tasks_service.GetTaskHandler(w, r, db)
-	case http.MethodPut:
-		tasks_service.EditTaskHandler(w, r, db)
-	case http.MethodDelete:
-		tasks_service.DeleteTaskHandler(w, r, db)
-	default:
-		http.Error(w, "Метод не поддерживается", http.StatusMethodNotAllowed)
 	}
 }
