--- tasks_service/update_task.go
@@ -1,55 +1,0 @@
-package tasks_service
-
-import (
-	"database/sql"
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"time"
-)
-
-// ОБНОВЛЕНИЕ ЗАДАЧИ
-func EditTaskHandler(w http.ResponseWriter, r *http.Request, db *sql.DB) {
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-
-	var task Task
-	if err := json.NewDecoder(r.Body).Decode(&task); err != nil {
-		http.Error(w, `{"error":"Ошибка декодирования JSON"}`, http.StatusBadRequest)
-		return
-	}
-	if task.ID == "" || task.Title == "" {
-		http.Error(w, `{"error":"Не указан id или заголовок"}`, http.StatusBadRequest)
-		return
-	}
-	tDate, err := time.Parse(dateFormat, task.Date)
-	if err != nil {
-		http.Error(w, `{"error":"Неверный формат даты"}`, http.StatusBadRequest)
-		return
-	}
-	now := time.Now()
-	if tDate.Before(now) {
-		if task.Repeat == "" {
-			task.Date = now.Format(dateFormat)
-		} else {
-			nextDate, err := NextDate(now, task.Date, task.Repeat)
-			if err != nil {
-				http.Error(w, fmt.Sprintf(`{"error":"%s"}`, err.Error()), http.StatusBadRequest)
-				return
-			}
-			task.Date = nextDate
-		}
-	}
-
-	query := `UPDATE scheduler SET date = ?, title = ?, comment = ?, repeat = ? WHERE id = ?`
-	res, err := db.Exec(query, task.Date, task.Title, task.Comment, task.Repeat, task.ID)
-	if err != nil {
-		http.Error(w, fmt.Sprintf(`{"error":"Ошибка обновления: %s"}`, err.Error()), http.StatusInternalServerError)
-		return
-	}
-	if rowsAffected, _ := res.RowsAffected(); rowsAffected == 0 {
-		http.Error(w, `{"error":"Задача не найдена"}`, http.StatusNotFound)
-		return
-	}
-	w.WriteHeader(http.StatusOK)
-	fmt.Fprintf(w, `{}`)
-}
